# src/main/resources/static/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Aluraflix",
    "version": "v1",
    "description": "API criada para o Challenge Alura Back-End"
  },
  "servers": [
    {
      "url": "https://lr-aluraflix.herokuapp.com"
    },
    {
      "url": "https://localhost:8080"
    }
  ],
  "paths": {
    "/api/videos/{id}": {
      "get": {
        "description": "Get Video By ID",
        "security": [
          {
            "BearerUserAuth": []
          }
        ],
        "tags": [
          "video-controller"
        ],
        "operationId": "getVideoById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseNotFound"
                }
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseUnauthorized"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Update Video By ID",
        "security": [
          {
            "BearerUserAuth": []
          }
        ],
        "tags": [
          "video-controller"
        ],
        "operationId": "updateVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseNotFound"
                }
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseUnauthorized"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Delete Video By ID",
        "security": [
          {
            "BearerUserAuth": []
          }
        ],
        "tags": [
          "video-controller"
        ],
        "operationId": "deleteVideo",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseSuccess"
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseNotFound"
                }
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/videos": {
      "get": {
        "description": "Get All Videos or Search For a Video",
        "security": [
          {
            "BearerUserAuth": []
          }
        ],
        "tags": [
          "video-controller"
        ],
        "operationId": "getVideos",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseNotFound"
                }
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseUnauthorized"
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Insert New Video",
        "security": [
          {
            "BearerUserAuth": []
          }
        ],
        "tags": [
          "video-controller"
        ],
        "operationId": "insertVideo",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponse"
                }
              }
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseUnauthorized"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerUserAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "BearerAdminAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "VideoRequest": {
        "required": [
          "description",
          "link",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "example": "Spring Boot acelere o desenvolvimento de microsserviços"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "example": "Vamos bater um papo sobre Spring Boot? O Fernando Boaglio, autor do livro, vai contar um pouco sobre sua experiência na área, projeto desenvolvido no livro e os pré-requisitos para uma boa leitura."
          },
          "link": {
            "type": "string",
            "example": "https://www.youtube.com/watch?v=TX1NztPbvMw"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32",
            "example": 2
          }
        }
      },
      "VideoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "link": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "MessageResponseSuccess": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Successfully!"
          }
        }
      },
      "MessageResponseNotFound": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "{resource} not found"
          }
        }
      },
      "MessageResponseUnauthorized": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2021-10-08T16:51:45.636+00:00"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Unauthorized"
          },
          "path": {
            "type": "string",
            "example": "/api/videos/1"
          }
        }
      }
    }
  }
}